Refact: combined all observation models into a single module.
